refactor(streaming): move SSE→StreamDelta mapping into anthropic-client
The Anthropic→StreamDelta conversion belongs in the provider client, not
the executor. Move sse_data_to_delta() into anthropic-client/convert.rs
and change send_prompt_streaming() to return StreamDelta directly. The
executor now just pipes deltas through without any Anthropic-specific
knowledge. SSE and streaming types are private again.
