+++ lib/anthropic-client/src/convert.rs
@@ -11,9 +11,12 @@ use llm::prompt::{
 };
 use llm::{PromptResponse, StopReason, Usage};
 
+use llm::stream::{ContentBlockType, StreamDelta};
+
 use crate::stream::{AccumulatedBlock, AccumulatedResponse, AccumulatedStopReason};
 use crate::types::{
-    AnthropicCacheControl, AnthropicContent, AnthropicMessage, AnthropicRequest,
+    AnthropicCacheControl, AnthropicContent, AnthropicContentBlock, AnthropicDelta,
+    AnthropicMessage, AnthropicRequest, AnthropicStopReason, AnthropicStreamEvent,
     AnthropicSystemBlock, AnthropicTool, AnthropicToolChoice, AnthropicToolResultContent,
 };
 
@@ -222,4 +225,87 @@ fn convert_accumulated_block(block: AccumulatedBlock) -> AssistantBlock {
             cache_control: None,
         },
     }
+}
+
+// ============================================================================
+// SSE JSON → StreamDelta
+// ============================================================================
+
+/// Parse a raw SSE data payload (Anthropic JSON) into a provider-agnostic
+/// [`StreamDelta`]. Returns `Ok(None)` for events that have no meaningful
+/// delta (e.g. `Ping`).
+pub(crate) fn sse_data_to_delta(data: &str) -> Result<Option<StreamDelta>, String> {
+    let event: AnthropicStreamEvent =
+        serde_json::from_str(data).map_err(|e| format!("JSON parse: {e}"))?;
+    Ok(match event {
+        AnthropicStreamEvent::MessageStart { message } => {
+            let input_tokens = message.usage.map(|u| u.input as u32).unwrap_or(0);
+            Some(StreamDelta::MessageStart { input_tokens })
+        }
+        AnthropicStreamEvent::ContentBlockStart {
+            index,
+            content_block,
+        } => {
+            let block_type = match content_block {
+                AnthropicContentBlock::Text => ContentBlockType::Text,
+                AnthropicContentBlock::ToolUse { id, name } => ContentBlockType::ToolUse {
+                    id: id.unwrap_or_default(),
+                    name: name.unwrap_or_default(),
+                },
+                AnthropicContentBlock::Thinking => ContentBlockType::Thinking,
+            };
+            Some(StreamDelta::ContentBlockStart {
+                index: index as usize,
+                block_type,
+            })
+        }
+        AnthropicStreamEvent::ContentBlockDelta { index, delta } => {
+            let idx = index as usize;
+            match delta {
+                AnthropicDelta::TextDelta { text } => text.map(|t| StreamDelta::TextDelta {
+                    index: idx,
+                    text: t,
+                }),
+                AnthropicDelta::ThinkingDelta { thinking } => {
+                    thinking.map(|t| StreamDelta::ThinkingDelta {
+                        index: idx,
+                        text: t,
+                    })
+                }
+                AnthropicDelta::InputJsonDelta { partial_json } => {
+                    partial_json.map(|j| StreamDelta::InputJsonDelta {
+                        index: idx,
+                        partial_json: j,
+                    })
+                }
+                AnthropicDelta::SignatureDelta { signature } => {
+                    signature.map(|s| StreamDelta::SignatureDelta {
+                        index: idx,
+                        signature: s,
+                    })
+                }
+            }
+        }
+        AnthropicStreamEvent::ContentBlockStop { index } => Some(StreamDelta::ContentBlockStop {
+            index: index as usize,
+        }),
+        AnthropicStreamEvent::MessageDelta { delta, usage } => {
+            let stop_reason = delta.stop_reason.map(|sr| match sr {
+                AnthropicStopReason::EndTurn => StopReason::EndTurn,
+                AnthropicStopReason::MaxTokens => StopReason::MaxTokens,
+                AnthropicStopReason::ToolUse => StopReason::ToolUse,
+                AnthropicStopReason::StopSequence => StopReason::StopSequence,
+            });
+            let output_tokens = usage.map(|u| u.output_tokens as u32).unwrap_or(0);
+            Some(StreamDelta::MessageDelta {
+                stop_reason,
+                output_tokens,
+            })
+        }
+        AnthropicStreamEvent::MessageStop => Some(StreamDelta::MessageStop),
+        AnthropicStreamEvent::Error { error } => Some(StreamDelta::Error {
+            message: error.message,
+        }),
+        AnthropicStreamEvent::Ping => None,
+    })
 }
